+++ __xstat64.c
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
 
+#ifdef __linux__
 #include <stdio.h>
 #include <fcntl.h>
 
@@ -11,7 +12,7 @@
 
 #include "iamroot.h"
 
-#ifdef __GLIBC__
+#ifdef _LARGEFILE64_SOURCE
 extern int __fxstatat64(int, int, const char *, struct stat64 *, int);
 
 int __xstat64(int ver, const char *path, struct stat64 *statbuf)
@@ -20,4 +21,5 @@ int __xstat64(int ver, const char *path, struct stat64 *statbuf)
 
 	return __fxstatat64(ver, AT_FDCWD, path, statbuf, 0);
 }
+#endif
 #endif