+++ fstat64.c
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
 
+#ifdef __linux__
 #include <stdio.h>
 #include <fcntl.h>
 
@@ -11,7 +12,7 @@
 
 #include "iamroot.h"
 
-#ifdef __GLIBC__
+#ifdef _LARGEFILE64_SOURCE
 int fstat64(int fd, struct stat64 *statbuf)
 {
 	__debug("%s(fd: %i, ...)\n", __func__, fd);
@@ -21,4 +22,5 @@ int fstat64(int fd, struct stat64 *statbuf)
 
 weak_alias(fstat64, __fstat64);
 weak_alias(fstat64, __fstat_time64);
+#endif
 #endif