+++ lstat64.c
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
 
+#ifdef __linux__
 #include <stdio.h>
 #include <fcntl.h>
 
@@ -11,7 +12,7 @@
 
 #include "iamroot.h"
 
-#ifdef __GLIBC__
+#ifdef _LARGEFILE64_SOURCE
 int lstat64(const char *path, struct stat64 *statbuf)
 {
 	__debug("%s(path: '%s', ...)\n", __func__, path);
@@ -21,4 +22,5 @@ int lstat64(const char *path, struct stat64 *statbuf)
 
 weak_alias(lstat64, __lstat64);
 weak_alias(lstat64, __lstat_time64);
+#endif
 #endif