+++ CHANGELOG.md
@@ -14,8 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Intercept the functions `acct()`, `catopen()`, `ftok()` and `mq_open()`
 - Intercept the Linux's functions `fanotify_mark()`, `inotify_add_watch()`,
   `swapon()` and `swapoff()`
-- Create the `64` weak aliases `fopen64()`, `freopen64()`, `open64()` and
-  `openat64()` for [musl] world
+- Create the `64` weak aliases `__fxstat64()`, `__fxstatat64()`,
+  `__lxstat64()`, `__xstat64()`, `fopen64()`, `freopen64()`, `fstat64()`,
+  `fstatat64()`, `lstat64()`, `open64()`, `openat64()` and `stat64()` for
+  [musl] world
 
 ### Changed